requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.14
Django==4.1.10
django-admin-interface==0.20.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-colorfield==0.9.0
django-cors-headers==4.0.0
django-enum-choices==2.1.4
django-environ==0.10.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
exceptiongroup==1.1.2
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==4.2.0
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
